Move kon core from config to util
